requirements.txt
numpy==1.23.1
moviepy==1.0.3
pafy==0.5.5
tqdm==4.64.0
google-api-python-client==2.55.0
random-word==1.0.7
ffmpeg-python==0.2.0